+++ src/main/java/io/github/jamalam360/jamlib/log/JamLibLogger.java
@@ -24,22 +24,24 @@
 
 package io.github.jamalam360.jamlib.log;
 
+import java.util.Arrays;
 import net.fabricmc.loader.api.FabricLoader;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * @author Jamalam
  */
 @SuppressWarnings("unused")
 public class JamLibLogger {
+
     private final String modId;
     private final boolean developmentOnly;
     private final Logger logger;
 
     private JamLibLogger(String modId, boolean developmentOnly) {
         this.modId = modId;
-        this.logger = LogManager.getLogger(modId);
+        this.logger = LoggerFactory.getLogger(modId);
         this.developmentOnly = developmentOnly;
     }
 
@@ -47,29 +49,31 @@ public class JamLibLogger {
         return new JamLibLogger(modId, false);
     }
 
-    public static JamLibLogger getDevelopmentOnlyLogger(String modId) { return new JamLibLogger(modId, true); }
+    public static JamLibLogger getDevelopmentOnlyLogger(String modId) {
+        return new JamLibLogger(modId, true);
+    }
 
-    private String addModId(String... messages) {
+    private String addModId(Object... messages) {
         if (!FabricLoader.getInstance().isDevelopmentEnvironment()) {
             messages[0] = "[" + modId + "] " + messages[0];
         }
 
-        return this.developmentOnly ? "(Development Only) " + String.join(" ", messages) : String.join(" ", messages);
+        return (this.developmentOnly ? "(Development Only) " : "") + String.join(" ", Arrays.stream(messages).map(Object::toString).toList());
     }
 
-    public void info(String... messages) {
+    public void info(Object... messages) {
         if (this.isActive()) {
             logger.info(addModId(messages));
         }
     }
 
-    public void warn(String... messages) {
+    public void warn(Object... messages) {
         if (this.isActive()) {
             logger.warn(addModId(messages));
         }
     }
 
-    public void error(String... messages) {
+    public void error(Object... messages) {
         if (this.isActive()) {
             logger.error(addModId(messages));
         }