+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - Automatically register entity attributes for entity types registered through JamLib.
 - Add `JamLibCompatibilityModuleHandler`.
 - `JamLibLogger` now takes varargs to it's logging methods.
+- Use SL4J rather than Log4J.
 
 [Full Changelog](https://github.com/JamCoreModding/jam-lib/compare/...)